+++ lib/bottommultinavigationbar.dart
@@ -1,5 +1,4 @@
 library bottommultinavigationbar;
 export 'package:bottommultinavigationbar/src/bottom_multi_navigation_bar.dart';
 export 'package:bottommultinavigationbar/src/page_navigator.dart';
-export 'package:bottommultinavigationbar/src/flutter_intent.dart';