+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.2.0] - 22.04.2020
+
+* added example (request #2)
+* removed flutter intent (outsourced as a separate package)
+* some improvements
+
 ## [1.1.2] - 09.02.2020
 
 * added new navigation system FlutterIntent for simplifying data exchange between widgets